requirements.txt
packaging==24.0
pillow==10.3.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
PyPDF2==3.0.1
pytesseract==0.3.10
tqdm==4.66.4
